requirements.txt
terminaltables~=3.1.10
python-dotenv~=1.0.1
requests~=2.32.3